FileTrait trait to ensure backend & cache support
